Move WieldAnimation to its own source files
